requirement.txt
absl-py==1.0.0
altair==4.2.0
anyio==3.4.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.3.1
astor==0.8.1
astunparse==1.6.3
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.1
backcall==0.2.0
base58==2.1.1
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
click==8.0.3
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
csc==20.2.1
cvlib==0.2.6
cycler==0.11.0
daphne==3.0.1
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.1
Django==3.0.6
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.13.1
drf-yasg==1.20.0
entrypoints==0.3
filelock==3.0.12
flatbuffers==2.0
fonttools==4.29.0
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.26
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
hyperlink==21.0.0
idna==2.10
imageio==2.14.1
importlib-metadata==4.10.1
imutils==0.5.4
incremental==21.3.0
inflection==0.5.1
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itypes==1.2.0
jedi==0.18.1
Jinja2==2.11.3
joblib==1.1.0
json5==0.9.6
jsonschema==4.3.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-server==1.13.1
jupyterlab==3.2.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.9.0
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lazy-object-proxy==1.5.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==1.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mnemonic==0.19
mysqlclient==2.0.3
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
numpy==1.20.1
oauthlib==3.1.1
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.0
progressbar==2.5
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.3
psycopg2==2.8.6
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.7.1
Pygments==2.10.0
Pympler==1.0.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pywin32==302
pywinpty==1.1.6
pyzmq==22.3.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.8
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
service-identity==18.1.0
six==1.15.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
sqlparse==0.4.1
streamlit==1.5.0
streamlit-lottie==0.0.3
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
uritemplate==3.0.1
urllib3==1.26.4
validators==0.18.2
virtualenv==20.4.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.12.1
zipp==3.7.0
zope.interface==5.2.0
